App Flow Change, Block Control Menu, Fixed SM
Now, when the user selects a block, the block control menu is shown first. This menu follows the camera! When the user clicks on the current placeholder button, the music selection menu will appear. After assigning a piece of music to the selected block, the user can select a different block and assign to it.
I also fixed the bug where the sounds were assigned wrong in the SoundManager.
POSSIBLE ISSUE: I just realized that, if the user switches modes while the block control menu is being displayed...it probably won't disappear. The same might be true for the music menu. We'll have to work out the logic for that later (unless the switch mode button is going to disappear -- we probably won't need it in production based on the flow of our code and the existence of things like back buttons?)