Fixed VFX and new selection issue
Deleted selection code before, then realized we still need it if we want to change a block's sound -- but then I realized the old code doesn't work anymore, so I came up with a quick and not-so-pretty work around (see MusicBlock and HomeManager).
I also fixed the VFX, which was only able to support data from a single sound before. Changed the data from static to something with instances!